requirements.txt
requests
py-cord
python_dotenv
aiohttp
BingImageCreator
openai
python-gemini-api
flask
psutil